pyproject.toml
[project]
name = "enrichment-agent"
version = "0.0.1"
description = "An agent that populates and enriches custom schemas"
authors = [
    { name = "Mohamed Ben Haddou", email = "mbenhaddou@mentis.io" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph==0.3.18",
    "langchain-openai==0.3.9",
    "langchain-anthropic>=0.1.23",
    "langchain==0.3.21",
    "langchain-fireworks>=0.1.7",
    "python-dotenv==1.0.1",
    "langchain-community==0.3.20",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.14",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.52.2",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cobble==0.1.4",
    "colourmap==1.1.21",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "d3graph==2.5.1",
    "dataclasses-json==0.6.7",
    "datazets==1.1.2",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "et-xmlfile==2.0.0",
    "faiss-cpu==1.11.0",
    "fonttools==4.57.0",
    "frozenlist==1.5.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "gotrue==2.11.4",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "ismember==1.0.6",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "joblib==1.5.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kiwisolver==1.4.8",
    "langchain-core==0.3.47",
    "langchain-text-splitters==0.3.7",
    "langgraph-checkpoint==2.0.21",
    "langgraph-prebuilt==0.1.4",
    "langgraph-sdk==0.1.58",
    "langsmith==0.3.18",
    "llvmlite==0.44.0",
    "lxml==6.0.0",
    "mammoth==1.9.1",
    "markupsafe==2.0.1",
    "marshmallow==3.26.1",
    "matplotlib==3.10.1",
    "msgpack==1.1.0",
    "multidict==6.2.0",
    "mypy-extensions==1.0.0",
    "narwhals==1.35.0",
    "networkx==3.4.2",
    "numba==0.61.2",
    "numpy==1.26.4",
    "openai==1.68.2",
    "openpyxl==3.1.5",
    "orjson==3.10.15",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "plotly==6.0.1",
    "postgrest==0.19.3",
    "propcache==0.3.0",
    "protobuf==5.29.4",
    "pyarrow==19.0.1",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "pydantic-core==2.27.2",
    "pydeck==0.9.1",
    "pynndescent==0.5.13",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-louvain==0.16",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "realtime==2.4.1",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.24.0",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.39",
    "storage3==0.11.3",
    "streamlit==1.44.1",
    "streamlit-d3graph==1.0.3",
    "strenum==0.4.15",
    "supabase==2.14.0",
    "supafunc==0.9.3",
    "tabulate==0.9.0",
    "tenacity==9.0.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "toml==0.10.2",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "typing-inspect==0.9.0",
    "typing-extensions==4.12.2",
    "tzdata==2025.2",
    "umap-learn==0.5.7",
    "urllib3==2.3.0",
    "vulture==2.14",
    "watchdog==6.0.0",
    "websockets==14.2",
    "yarl==1.18.3",
    "zstandard==0.23.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest-asyncio"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["enrichment"]
[tool.setuptools.package-dir]
"enrichment" = "./enrichment"
"langgraph.templates.enrichment" = "./enrichment"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
include = ["*.py", "*.pyi", "*.ipynb"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
"ntbk/*" = ["D", "UP", "T201"]
[tool.ruff.lint.pydocstyle]
convention = "google"
